feat: Add top content headings and paragraphs
In this commit, the headings and paragraphs text of the top content of the main
content section of the web page is update using the provided headings and paragraphs.
to achieve this, first we select all the h4 headings of the within the div of class
'.text-contenet'. the array-like result is then traversed using numeric index to
target the features and the about h4 headings.

a similar approach is used for the paragraphs within the div with class '.text-content'.

document.querySelectorAll was preferably used as it returns all the h4 and p within
the targeted div class, the result of which can be traversed and updated accordingly