feat: Add bottom content of web page
In this commit, using the DOM, I updated the heading and the paragraph of the bottom
content. the div containing the bottom h4 and p tags also has the class of .text-content
which was previous selected using querySelectorAll('.text-content'). I simply traversed
the result of the previous querySelectorAll('.text-content') and selectively update the
h4 and p content by passing the correct index