feat: Update the contact section of web page
In this commit, using the DOM, I updated the h4 and paragraphs of the contact section
of the web page. to achieve this, using querySelector, I selected the the h4 within
the class '.contact' and updated its textContent prop using the provided data.

Similarly, the three other paragraphs present within the div of class '.contact' were
also selected using document.querySelectorAll and their content was updated using
relevant data supplied